Move default main implementation header to hpx_init
