+++ init/src/main.cpp
@@ -5,7 +5,6 @@
 //  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 
 #include <hpx/config.hpp>
-#include <hpx/wrap/config/defines.hpp>
 
 // We support different implementation depending upon the Operating
 // System in use.
@@ -15,7 +14,7 @@
 
 // when statically linking the default definition of main happens on demand only
 #if !defined(HPX_HAVE_STATIC_LINKING)
-#include <hpx/wrap_main_impl.hpp>
+#include <hpx/hpx_main_impl.hpp>
 #endif
 
 #endif
