+++ wrap/CMakeLists.txt
@@ -6,50 +6,17 @@
 # Distributed under the Boost Software License, Version 1.0. (See accompanying
 # file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 
-set(HPX_WITH_DYNAMIC_HPX_MAIN_DEFAULT OFF)
-if("${CMAKE_SYSTEM_NAME}" STREQUAL "Linux" OR APPLE)
-  set(HPX_WITH_DYNAMIC_HPX_MAIN_DEFAULT ON)
-endif()
-
-hpx_option(
-  HPX_WITH_DYNAMIC_HPX_MAIN
-  BOOL
-  "Enable dynamic overload of system ``main()`` (Linux and Apple only, default: ON)"
-  ${HPX_WITH_DYNAMIC_HPX_MAIN_DEFAULT}
-  ADVANCED
-)
-if(HPX_WITH_DYNAMIC_HPX_MAIN)
-  if(NOT HPX_WITH_DYNAMIC_HPX_MAIN_DEFAULT)
-    hpx_error(
-      "HPX_WITH_DYNAMIC_HPX_MAIN was set to ON, but the option is only available on Linux and Apple (this is \"${CMAKE_SYSTEM_NAME}\")."
-    )
-  endif()
-  hpx_add_config_define_namespace(
-    DEFINE HPX_HAVE_DYNAMIC_HPX_MAIN NAMESPACE wrap
-  )
-endif()
-
-set(config_header
-    "${CMAKE_CURRENT_BINARY_DIR}/include/hpx/wrap/config/defines.hpp"
-)
-write_config_defines_file(NAMESPACE wrap FILENAME ${config_header})
-
 set(hpx_wrap_HEADERS
     ""
     CACHE INTERNAL "Headers for libhpx_wrap." FORCE
 )
-list(APPEND hpx_wrap_HEADERS hpx/hpx_main.hpp hpx/wrap_main.hpp
-     hpx/wrap_main_impl.hpp
-)
+list(APPEND hpx_wrap_HEADERS hpx/hpx_main.hpp hpx/wrap_main.hpp)
 
 set(hpx_wrap_SOURCES
     ""
     CACHE INTERNAL "Sources for libhpx_wrap." FORCE
 )
-list(APPEND hpx_wrap_SOURCES main.cpp)
-if(HPX_WITH_DYNAMIC_HPX_MAIN)
-  list(APPEND hpx_wrap_SOURCES hpx_wrap.cpp)
-endif()
+list(APPEND hpx_wrap_SOURCES hpx_wrap.cpp)
 
 # make source groups
 set(hpx_wrap_SOURCE_ROOT ${CMAKE_CURRENT_SOURCE_DIR}/src)
@@ -69,12 +36,6 @@ add_hpx_source_group(
   ROOT "${hpx_wrap_HEADER_ROOT}"
   TARGETS ${hpx_wrap_HEADERS}
 )
-add_hpx_source_group(
-  NAME hpx_wrap
-  CLASS "Generated Files"
-  ROOT "${CMAKE_CURRENT_BINARY_DIR}/include/hpx"
-  TARGETS ${config_header}
-)
 
 add_library(
   hpx_wrap STATIC ${hpx_wrap_SOURCES} ${hpx_wrap_HEADERS} ${config_header}
@@ -85,10 +46,8 @@ target_link_libraries(hpx_wrap PRIVATE hpx_init)
 target_link_libraries(hpx_wrap PRIVATE hpx_private_flags)
 target_compile_definitions(hpx_wrap PRIVATE HPX_APPLICATION_EXPORTS)
 target_include_directories(
-  hpx_wrap
-  PUBLIC $<BUILD_INTERFACE:${hpx_wrap_HEADER_ROOT}>
-         $<BUILD_INTERFACE:${CMAKE_CURRENT_BINARY_DIR}/include>
-         $<INSTALL_INTERFACE:include>
+  hpx_wrap PUBLIC $<BUILD_INTERFACE:${hpx_wrap_HEADER_ROOT}>
+                  $<INSTALL_INTERFACE:include>
 )
 
 set_property(TARGET hpx_wrap PROPERTY FOLDER "Core")
@@ -125,12 +84,6 @@ install(
 
 install(
   DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR}/include/
-  DESTINATION ${CMAKE_INSTALL_INCLUDEDIR}
-  COMPONENT runtime
-)
-
-install(
-  DIRECTORY ${CMAKE_CURRENT_BINARY_DIR}/include/hpx
   DESTINATION ${CMAKE_INSTALL_INCLUDEDIR}
   COMPONENT runtime
 )
